Moved files for new repository directory structure.
