--- APaRSer.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "APaRSer", "APaRSer.csproj", "{BCA572AC-AF0A-40B6-AD29-795841A65CE2}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "APaRSerUnitTests", "APaRSerUnitTests\APaRSerUnitTests.csproj", "{8BCF50EE-C4F7-45D0-9D1B-9018E8C33B05}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{BCA572AC-AF0A-40B6-AD29-795841A65CE2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BCA572AC-AF0A-40B6-AD29-795841A65CE2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BCA572AC-AF0A-40B6-AD29-795841A65CE2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BCA572AC-AF0A-40B6-AD29-795841A65CE2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8BCF50EE-C4F7-45D0-9D1B-9018E8C33B05}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8BCF50EE-C4F7-45D0-9D1B-9018E8C33B05}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8BCF50EE-C4F7-45D0-9D1B-9018E8C33B05}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8BCF50EE-C4F7-45D0-9D1B-9018E8C33B05}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
