--- README.md
@@ -1,3 +1,0 @@
-APaRSer: an APRS packat parser, encoder, and decoder written in C# by Cameron Bielstein (KG5GHQ).
-
-The VS solution has been written as a Xamarin Portable Class Library to support .NET Framework, Windows, Windows Phone, Android, and iOS, for cross platform compatibility.
