+++ src/APaRSer/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="GeoCoordinate" version="1.1.0" targetFramework="portable45-net45+win8+wpa81" />
+</packages>
