+++ src/APaRSer/packages/GeoCoordinate.1.1.0/LICENSE.md
